pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "design-from-url"
version = "0.1.0"
description = "Extract design tokens (colors, typography, spacing) from a URL and emit a DESIGN.md file conformant to @google/design.md spec."
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = { text = "MIT" }
authors = [{ name = "Xiaofei Ge" }]
# Runtime browser is provided by `agent-browser` (external CLI on PATH); see
# README. Pillow for Phase 1.5b screenshot pixel sampling; PyYAML for Phase 2.5
# schema_fixer parse side (write side stays on custom yaml_emit.py for quoting).
dependencies = [
    "pillow>=12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
design-from-url = "design_from_url.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/design_from_url"]

[tool.design-from-url]
design_md_npm_version = "0.1.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
